(v2.2.1) - Gymnasium compability (#280)
* Updated sinergym imports to gymnasium package

* Updated reset return to new gymnasium format

* Added seed and option parameters in reset method (according to gymnasium standard)

* Updated step method parameters for including new gymnasium standard

* weather variability included in reset options (options by default functionality included too)

* wrappers adapted to gymnasium format

* Rename done parameter to terminated in logger

* Temporaly disabled stable baselines 3 from extra requires due to gymnasium incompability (warning in README.md)

* Fixed warning advise to get thread name in simulator

* Fixed test for controllers with gymnasium

* Adapted env tests to gymnasium

* Fixed env_checker and update code to get gymnasium environments id's in environment test

* Fixed reward tests to gymnasium format

* Fixed simulator tests to gymnasium format and some changes about new info output in that layer

* Rename done argument to terminated (gymnasium name) in LoggerWrapper

* Fixed wrappers

* Fixed action space bounds casting to float32 type in order to avoid gymnasium warnings

* Fixed render mode and action space dtype bounds in order to avoid gymnasium warnings

* Fixed bug with MultiObs wrappers (parameter keys required but I don't know why)

* Updated notebooks

* Updated scripts

* Updated version from 2.2.0 to 2.2.1

* Fixed pytype errors

+++ sinergym/utils/constants.py
@@ -2,7 +2,7 @@
 
 import os
 
-import gym
+import gymnasium as gym
 import numpy as np
 import pkg_resources
 
@@ -373,8 +373,8 @@ DEFAULT_5ZONE_ACTION_MAPPING = {
 DEFAULT_5ZONE_ACTION_SPACE_DISCRETE = gym.spaces.Discrete(10)
 
 DEFAULT_5ZONE_ACTION_SPACE_CONTINUOUS = gym.spaces.Box(
-    low=np.array([15.0, 22.5]),
-    high=np.array([22.5, 30.0]),
+    low=np.array([15.0, 22.5], dtype=np.float32),
+    high=np.array([22.5, 30.0], dtype=np.float32),
     shape=(2,),
     dtype=np.float32
 )
@@ -440,8 +440,8 @@ DEFAULT_DATACENTER_ACTION_MAPPING = {
 DEFAULT_DATACENTER_ACTION_SPACE_DISCRETE = gym.spaces.Discrete(10)
 
 DEFAULT_DATACENTER_ACTION_SPACE_CONTINUOUS = gym.spaces.Box(
-    low=np.array([15.0, 22.5]),
-    high=np.array([22.5, 30.0]),
+    low=np.array([15.0, 22.5], dtype=np.float32),
+    high=np.array([22.5, 30.0], dtype=np.float32),
     shape=(2,),
     dtype=np.float32)
 
@@ -503,8 +503,8 @@ DEFAULT_WAREHOUSE_ACTION_MAPPING = {
 DEFAULT_WAREHOUSE_ACTION_SPACE_DISCRETE = gym.spaces.Discrete(10)
 
 DEFAULT_WAREHOUSE_ACTION_SPACE_CONTINUOUS = gym.spaces.Box(
-    low=np.array([15.0, 22.5, 15.0, 22.5, 15.0]),
-    high=np.array([22.5, 30.0, 22.5, 30.0, 22.5]),
+    low=np.array([15.0, 22.5, 15.0, 22.5, 15.0], dtype=np.float32),
+    high=np.array([22.5, 30.0, 22.5, 30.0, 22.5], dtype=np.float32),
     shape=(5,),
     dtype=np.float32)
 
@@ -580,8 +580,8 @@ DEFAULT_OFFICE_ACTION_MAPPING = {
 DEFAULT_OFFICE_ACTION_SPACE_DISCRETE = gym.spaces.Discrete(10)
 
 DEFAULT_OFFICE_ACTION_SPACE_CONTINUOUS = gym.spaces.Box(
-    low=np.array([15.0, 22.5]),
-    high=np.array([22.5, 30.0]),
+    low=np.array([15.0, 22.5], dtype=np.float32),
+    high=np.array([22.5, 30.0], dtype=np.float32),
     shape=(2,),
     dtype=np.float32)
 
@@ -690,8 +690,8 @@ DEFAULT_OFFICEGRID_ACTION_MAPPING = {
 DEFAULT_OFFICEGRID_ACTION_SPACE_DISCRETE = gym.spaces.Discrete(10)
 
 DEFAULT_OFFICEGRID_ACTION_SPACE_CONTINUOUS = gym.spaces.Box(
-    low=np.array([15.0, 22.5, 0.0, 0.0]),
-    high=np.array([22.5, 30.0, 1.0, 1.0]),
+    low=np.array([15.0, 22.5, 0.0, 0.0], dtype=np.float32),
+    high=np.array([22.5, 30.0, 1.0, 1.0], dtype=np.float32),
     shape=(4,),
     dtype=np.float32
 )
@@ -767,8 +767,8 @@ DEFAULT_SHOP_ACTION_MAPPING = {
 DEFAULT_SHOP_ACTION_SPACE_DISCRETE = gym.spaces.Discrete(10)
 
 DEFAULT_SHOP_ACTION_SPACE_CONTINUOUS = gym.spaces.Box(
-    low=np.array([15.0, 22.5]),
-    high=np.array([22.5, 30.0]),
+    low=np.array([15.0, 22.5], dtype=np.float32),
+    high=np.array([22.5, 30.0], dtype=np.float32),
     shape=(2,),
     dtype=np.float32
 )
@@ -812,8 +812,8 @@ DEFAULT_SHOP_ACTION_DEFINITION = {
 # DEFAULT_HOSPITAL_ACTION_SPACE_DISCRETE = gym.spaces.Discrete(10)
 
 # DEFAULT_HOSPITAL_ACTION_SPACE_CONTINUOUS = gym.spaces.Box(
-#     low=np.array([15.0, 22.5]),
-#     high=np.array([22.5, 30.0]),
+#     low=np.array([15.0, 22.5], dtype=np.float32),
+#     high=np.array([22.5, 30.0], dtype=np.float32),
 #     shape=(2,),
 #     dtype=np.float32)
